move entities test to the generator directory
