--- generators/entities/generator.spec.mjs
@@ -1,44 +1,0 @@
-/**
- * Copyright 2013-2022 the original author or authors from the JHipster project.
- *
- * This file is part of the JHipster project, see https://www.jhipster.tech/
- * for more information.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-import { jestExpect as expect } from 'mocha-expect-snapshot';
-import lodash from 'lodash';
-import { basename, dirname } from 'path';
-import { fileURLToPath } from 'url';
-
-import testSupport from '../../test/support/index.cjs';
-import Generator from './index.js';
-
-const { snakeCase } = lodash;
-const { testBlueprintSupport } = testSupport;
-
-const __filename = fileURLToPath(import.meta.url);
-const __dirname = dirname(__filename);
-
-const generator = basename(__dirname);
-
-describe(`JHipster ${generator} generator`, () => {
-  it('generator-list constant matches folder name', async () => {
-    await expect((await import('../generator-list.js')).default[`GENERATOR_${snakeCase(generator).toUpperCase()}`]).toBe(generator);
-  });
-  it('should support features parameter', () => {
-    const instance = new Generator([], { help: true }, { bar: true });
-    expect(instance.features.bar).toBe(true);
-  });
-  describe.skip('blueprint support', () => testBlueprintSupport(generator));
-});
